Adding a compile time check that base class should be polymorphic too if a derived class is polymorphic.
If derived class is polymorphic it probably contains a v-table inside. Thus,
pointer to `this` may have an offset to instance data in member functions.
In order to call methods in base class properly with pointers to those member
functions, the base class also should have v-table and consequentially, be
polymorphic too.