Fixing possible memory leaks in get_external_data/set_external_data functions
If external data size is more than size of pointer, a copy of the data is allocated and stored in the V8 External value. To delete this copy of data, we are making the External value as a Persistent weak handle. When this persistent handle would be garbage collected, the data copy would be deleted in the a delete_ext_data() callback function.