Cast pointers to class members in `class_::set` functions
A `class_` and `property_` allows to bind pointers to base members,
that is fine. But on unwrapping the base class type is used, which may be
not registered in v8pp internal structures. Such pointers to base members shall
be converted to the actual derived class members.

Added `pointer_type` alias into `detail::function_traits` template specializations
for class members. Perform a cast to the class pointer type in the `class_::set`
functions to unwrap objects of this class type later on members invocation.

Added tests for `function_traits` and `class_`. Inherit a test class `X` from
another `Xbase` and bind the inherited members in `X` class binding.