Eliminate `class_singleton` template, move all functions to base `class_info`
Make this `class_info` moveable-only and store it in `std:list` to ensure that
all pointers and references of `class_info` instances are still valid after
`class_singletons::add_class()/remove_class()` invocations.

Removed unnecessary test for inheriting in other isolate, because if there
is no such a base class in that isolate, an exception would be thrown,
otherwise it is absolutely legal.

Removed outdated weak callback signature with `v8::WeakCallbackData` type.

# Conflicts:
#	test/test_class.cpp
#	v8pp/class.hpp