add a test case for `JSON.stringify()` with wrapped class (issue #88)
According to the [`JSON.stringify()` documentation][1] it's possible to add a
function `toJSON()` to prototype for custom conversion to JSON.

Added a `Xbase::to_json()` functions and exposed it as `toJSON()` into v8,
so that `JSON.stringify()` call for such objects would invoke the conversion
function.

Updated `create_X()` factory function to set the `X::var` on construction at
JavaScript side.

[1]: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/JSON/stringify#toJSON()_behavior