Fix compile error when using C++14 mode with gcc
On linux, v8pp failed to compile in it's default configuration defined in the
makefile (C++14 mode). When compiling in C++14 mode, `__has_include(<string_view>)`
returns true and also the header can be included. However the std::string_view is
not defined. This uses the macro `__cpp_lib_string_view` instead to check for
`std::string_view` availability. It compiles now in both C++14 and C++17 mode.