Proper implementation for move ctor and assignment in `context`
Extract context destruction into `destroy()` helper to call it in the move assigned.
Reset `isolate_` and `impl_` members to nullptr in the `context::destroy()`
to have the context in a valid moved out state.

Add explicit move/copy ctor and assignment operators for `context`, `module`, `class_`,
and static_assert for them into the tests.

Update test cases for move ctor and assignment in test_context.cpp